% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AB_qPCR_export.to.MeltR.csv.R
\name{AB.qPCR.export.to.MeltR.csv}
\alias{AB.qPCR.export.to.MeltR.csv}
\title{Reformats an Applied Biosystems qPCR export file for fitting with MeltR}
\usage{
AB.qPCR.export.to.MeltR.csv(
  data_file,
  index_file,
  output_name = NA,
  Temp_correction = NA,
  Scale = NA
)
}
\arguments{
\item{data_file}{Path to the tab delimited AB raw data export file.}

\item{index_file}{Path to the .csv formatted index file.}

\item{output_name}{Optional name of the .csv file you want to write. Default = NA to not write a file.}

\item{Temp_correction}{Optional function to apply to correct the temperature readings, determined by callibrating the temperature in the qPCR machine. Default = NA to not apply a correction.}

\item{Scale}{Option to scale the emission value by dividing by a value specified here. Default = NA not to scale the Emission values.}
}
\value{
A data frame containing the Normalized emission and temperature data for the wells you specified with the index file.
}
\description{
This function reads an Applied Biosystems (AB) qPCR export file (Text tab delimited), pulls out Temperature and normalized emission data,
and formats it into a R data frame that MeltR can fit. Requires a .csv file index describing what reagents
are in each well. The index should have columns labeled "well", "A" for the first reagent,
and "B" for the second reagent. Also an option to save the result as a .csv file (see output_name)
and to apply a Temperature correction function.
}
