% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLTrimmer.R
\name{BLTrimmer}
\alias{BLTrimmer}
\title{Auto-trim absorbance melting curves to obtain more accurate meltR.A fits}
\usage{
BLTrimmer(
  meltR.A.fit,
  n.combinations = 100,
  n.ranges = 5,
  range.step = 5,
  no.trim.range = c(0.2, 0.8),
  parallel = "none",
  n.core = 1,
  Save_results = "none",
  file_path = getwd(),
  file_prefix = "BLTrimmer"
)
}
\arguments{
\item{meltR.A.fit}{data_frame containing absorbance melting data}

\item{n.combinations}{The blank sample for background subtraction, or a list of blanks to apply to different samples for background subtraction. "none" to turn off background subtraction. If there is a single blank in the data set, The identity of the blank, for example, blank = 1 or blanke = "blank". If there are multiple blanks in the data, blank = list(c("blank 1", "Sample 1"), c("blank 2", "sample 2")) and so on. Sample identifiers should be what they are in the data frame. If you need to figure out what the sample identifiers are, run unique(df$Sample), where df is the name of the R data frame you are using, in your R console.}

\item{n.ranges}{A vector containing the Nucleic acid type and the sequences you are fitting.}

\item{range.step}{The molecular model you want to fit. Options: "Monomolecular.2State", "Monomolecular.3State", "Heteroduplex.2State", "Homoduplex.2State".}

\item{no.trim.range}{The thermodynamic model you want to fit. Options: "VantHoff". Default = "VantHoff".}

\item{parallel}{The temperature used to calculate the NucAcid concentration. Default = 90.}

\item{n.core}{Option to only fit certain temperature ranges for melting curves. Either a vector or a list. If this is set to a vector, meltR.A will only fit temperatures in this range for all melting curves Example = c(17, 75). If set to a list of vectors, meltR.A will change what values are fit for each curve. Example, list(c(0,100), c(17,75), .... , c(0,100)). The length of this list has to be the equal to the number of samples that will be fit.}

\item{Save_results}{what methods do you want to use to fit data. Default = c(TRUE, TRUE, TRUE). Can be true or false. Note, method 1 must be set to TRUE or the subsequent steps will not work.}

\item{file_path}{What results to save. Options: "all" to save PDF plots and ".csv" formated tables of parameters, "some" to save ".csv" formated tables of parameters, or "none" to save nothing.}

\item{file_prefix}{Prefix that you want on the saved files.}
}
\value{
A list of data frames containing parameters from the fits and data for plotting the results with ggplot2.
}
\description{
Automates #'
}
