% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{clipboard}
\alias{clipboard}
\title{Returns values from the clipboard}
\usage{
clipboard()
}
\description{
This is useful in the \code{\link[readr:read_delim]{readr::read_delim()}} functions to read from the clipboard.
}
\examples{
\dontrun{
  clipboard()
}
}
\seealso{
readr::read_delim
}
