% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.abs.data}
\alias{df.abs.data}
\title{df.abs.data}
\format{
A data frame with 1800 rows and 4 variables:
\describe{
\item{\code{Sample}}{double The sample melted in different cuvettes. Sample 1 is a buffer blank and other samples contain different concentrations of RNA}
\item{\code{Pathlength}}{double The pathlength of the cuvette for each sample in cm}
\item{\code{Temperature}}{double The temperature where the absorbance was recorded, in degrees Celcius}
\item{\code{Absorbance}}{double The absorbance at 260 nm}
}
}
\usage{
df.abs.data
}
\description{
Absorbance melting curves for a heteroduplex helix composed of CGAAAGGU/ACCUUUCG. Samples contain strands at equamolar concentrations in 1 M NaCl 20 mM MOPS 0.01 mM EDTA 0.001\% SDS pH 7.0.
}
\keyword{datasets}
