% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.BLtrimmer.test.results}
\alias{df.BLtrimmer.test.results}
\title{df.BLtrimmer.test.results}
\format{
A data frame with 90 rows and 5 variables:
\describe{
\item{\code{Method}}{The meltR.A method}
\item{\code{dH}}{The enthalpy in kcal/mol}
\item{\code{dS}}{The entropy in cal/mol/K}
\item{\code{dG}}{The Gibbs free energy at 37C in kcal/mol}
\item{\code{Tm_at_0.1mM}}{An expectation maximized Tm at Ct = 0.1 mM in Celcius}
}
}
\usage{
df.BLtrimmer.test.results
}
\description{
The expected results from the testing protocol if everything is working.
}
\keyword{datasets}
