% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.ext.data}
\alias{df.ext.data}
\title{df.ext.data}
\format{
A data frame with 323 rows and 5 variables:
\describe{
\item{\code{Wavelength.nm}}{The wavelength that the extinction coefficient will be calculated}
\item{\code{Absorbtivity.mM}}{The absorbtivity in mM for a nucleotide}
\item{\code{Absorbtivity.M}}{The absorbtivity in M for a nucleotide}
\item{\code{Nucleotide}}{The nucleotide identity}
\item{\code{Reference}}{The reference the absorbtivity values come from}
}
}
\usage{
df.ext.data
}
\description{
Data for calculating extinction coefficients from DNA and RNA sequences
}
\keyword{datasets}
