% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.FAM.C.BHQ1.data}
\alias{df.FAM.C.BHQ1.data}
\title{df.FAM.C.BHQ1.data}
\format{
A data frame with 1246 rows and 4 variables:
\describe{
\item{\code{Sample}}{ The sample melted in different cuvettes. Sample 1 is a buffer blank and other samples contain different concentrations of RNA}
\item{\code{Pathlength}}{ The pathlength of the cuvette for each sample in cm}
\item{\code{Temperature}}{ The temperature where the absorbance was recorded, in degrees Celcius}
\item{\code{Absorbance}}{ The absorbance at 260 nm}
}
}
\usage{
df.FAM.C.BHQ1.data
}
\description{
Absorbance melting curves for a heteroduplex helix composed of FAM-CUGAGUC/GACUCAG-BHQ1. Samples contain strands at equamolar concentrations in 1 M NaCl 20 mM MOPS 0.01 mM EDTA 0.001\% SDS pH 7.0. The data has been background subtracted.
}
\keyword{datasets}
