% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.test.data}
\alias{df.test.data}
\title{df.test.data}
\format{
A data frame with 14661 rows and 5 variables:
\describe{
\item{\code{Sample}}{ The sample melted in different cuvettes. Sample 1 is a buffer blank and other samples contain different concentrations of RNA}
\item{\code{Temperature}}{ The temperature where the absorbance was recorded, in degrees Celcius}
\item{\code{Pathlength}}{ The pathlength of the cuvette for each sample in cm}
\item{\code{Absorbance}}{ The absorbance at 260 nm}
\item{\code{DS}}{ The data set. DS1: A monomolecular 2 state helix with no blanks.  DS2: A Heteroduplex 2 state helix with no blanks. DS3: A Homoduplex 2 state helix with no blanks. DS4: A Monomolecular 2 state helix with 1 blank. DS5: A Heteroduplex 2 state helix with 1 blank. DS6: A Homoduplex 2 state helix with 1 blank. DS7: A Monomolecular 2 state helix with 3 blanks. DS8: A Heteroduplex 2 state helix with 3 blanks.  DS9: A Homoduplex 2 state helix with 3 blanks.}
}
}
\usage{
df.test.data
}
\description{
Modeled absorbance melting curves for a data sets used in the testing protocols.
}
\keyword{datasets}
