% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.test.results}
\alias{df.test.results}
\title{df.test.results}
\format{
A data frame with 10368 rows and 9 variables:
\describe{
\item{\code{Method}}{The meltR.A method}
\item{\code{dH}}{The enthalpy in kcal/mol}
\item{\code{SE.dH}}{The error in the enthalpy in kcal/mol}
\item{\code{dS}}{The entropy in cal/mol/K}
\item{\code{SE.dS}}{The error in the entropy in cal/mol/K}
\item{\code{dG}}{The Gibbs free energy at 37C in kcal/mol}
\item{\code{SE.dG}}{The error in the Gibbs free energy at 37C in kcal/mol}
\item{\code{Tm_at_0.1mM}}{An expectation maximized Tm at Ct = 0.1 mM in Celcius}
\item{\code{SE.Tm_at_0.1mM}}{The error in the expectation maximized Tm at Ct = 0.1 mM in Celcius}
}
}
\usage{
df.test.results
}
\description{
The expected results from the testing protocol if everything is working.
}
\keyword{datasets}
