% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.states.R
\name{enumerate.states}
\alias{enumerate.states}
\title{Enumerates all the possible base pairing states for a RNA helix and compiles a list of thermodynamic parameters for each state}
\usage{
enumerate.states(NucAcid = c("RNA", "F-CGAAAGGU", "ACCUUUCG-Q"))
}
\arguments{
\item{NucAcid}{A vector containing RNA or DNA sequences. Flourophores and quenchers should be separated from the RNA sequnece with a dash. Example "c("RNA", "FAM-ACCUUUCG", "CGAAAGGU-BHQ1")"}
}
\value{
A list of data frames containing the sequence of the A strand, the sequence of the B strand, and a 1 if a BP is present or a 0 if no BP is present.e
}
\description{
Generates all possible base pairing states for a helix. Assumes that strands are not allowed to slips.
}
