% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helix.energy.R
\name{Helix.energy}
\alias{Helix.energy}
\title{Calculates folding INN-HB parameters for a helix sequnece}
\usage{
Helix.energy(
  seqF = "ACGCA",
  seqR = "UGCGU",
  output = "df",
  F.Q = FALSE,
  Double.label = 0,
  Fluor = "FAM",
  Qunecher = "BHQ1",
  AA.UU = -0.9,
  AU.AU = -1.1,
  UA.UA = -1.3,
  CU.AG = -2.1,
  CA.UG = -2.1,
  GU.AC = -2.2,
  GA.UC = -2.1,
  CG.CG = -2.4,
  GG.CC = -3.3,
  GC.GC = -3.4,
  Initiation = 4.09,
  Term.AU = 0.45,
  FAMC.GBHQ1 = -3.93,
  FAMU.ABHQ1 = -3,
  FAMG.CBHQ1 = -4,
  FAMA.UBHQ1 = -2.5,
  Symmetry = 0.43
)
}
\arguments{
\item{seqF}{Forward RNA sequence}

\item{seqR}{Reverse RNA sequence}

\item{output}{The format you want the data to be in. "df" for a dataframe counting the parameters or "energy" for the energy of folding.}

\item{F.Q}{Is there a fluorophore or a quencher on the helix.}

\item{AA.UU}{= -0.9, #UU.AA 1 Energy for a folding parameter}

\item{AU.AU}{= -1.1,#AU.AU Energy for a folding parameter}

\item{UA.UA}{= -1.3, #UA.UA Energy for a folding parameter}

\item{CU.AG}{= -2.1, #AG.CU 2 Energy for a folding parameter}

\item{CA.UG}{= -2.1, #UG.CA 3 Energy for a folding parameter}

\item{GU.AC}{= -2.2, #AC.GU 4 Energy for a folding parameter}

\item{GA.UC}{= -2.1, #UC.GA 5 Energy for a folding parameter}

\item{CG.CG}{= -2.4, #CG.CG Energy for a folding parameter}

\item{GG.CC}{= -3.3, #CC.GG 6 Energy for a folding parameter}

\item{GC.GC}{= -3.4, #GC.GC Energy for a folding parameter}

\item{Initiation}{= 4.09 Energy for a folding parameter}

\item{Term.AU}{= 0.45 Energy for a folding parameter}

\item{FAMC.GBHQ1}{= -3.93 Energy for a folding parameter}

\item{Symmetry}{= 0.43 Energy for a folding parameter}
}
\value{
A data frame containing the data in the "information in the ".o3a" formatted file.
}
\description{
Calculates folding INN-HB parameters for a helix sequnece
}
