% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.A.derivative.R
\name{meltR.A.derivative}
\alias{meltR.A.derivative}
\title{Takes the first and second derivatives of a melting curves}
\usage{
meltR.A.derivative(data_frame, n = 4)
}
\arguments{
\item{data_frame}{The MeltR data frame you want to take the derivative of. Columns must be named "Sample", "Pathlength", "Temperature", and "Absorbance.}

\item{n}{The number of data point you want to smooth the first derivative to. Default n = 4.}
}
\value{
A list of data frames. One data frame contains the first derivative and one data frame contains the second derivative.
}
\description{
Takes the first and second derivatices of all the melting curves in a meltR data frame. Returns a list of dataframes, one containing the first deeivative and one containing the second.
}
