% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.A.model.R
\name{meltR.A.model}
\alias{meltR.A.model}
\title{Models realistic absorbance melting curves}
\usage{
meltR.A.model(
  NucAcid = c("RNA", "CGCGCG"),
  Samples = 9,
  Absorbance_error = 0.01,
  Mmodel = "Homoduplex.2State",
  Tmodel = "VantHoff"
)
}
\arguments{
\item{NucAcid}{A vector containing the Nucleic acid type and the sequences you are fitting for calculating extinction coefficients. Examples: c("RNA", "UUUUUU", "AAAAAA"), c("DNA", "GCTAGC"), etc... . For a custom extinction coefficient enter "Custom" followed by the molar extinction coefficients for every nucleic acid in the sample. For example, c("Custom", 10000, 20000).}

\item{Samples}{The number of samples you want to model}

\item{Absorbance_error}{The amount of absorbance error}

\item{Mmodel}{The molecular model you want to fit. Options: "Monomolecular.2State", "Monomolecular.3State", "Heteroduplex.2State", "Homoduplex.2State".}

\item{Tmodel}{The thermodynamic model you want to fit. Options: "VantHoff". Default = "VantHoff".}
}
\value{
A data frame containing modeled data
}
\description{
Models realistic absorbance melting curves
}
