% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.a.nlme.R
\name{meltR.A.nlme}
\alias{meltR.A.nlme}
\title{Experimental code: not ready for use}
\usage{
meltR.A.nlme(
  data_frame,
  blank,
  NucAcid,
  concT = 90,
  Mmodel,
  Tmodel = "VantHoff",
  Save_results = "none",
  file_prefix = "Fit",
  file_path = getwd()
)
}
\arguments{
\item{data_frame}{data_frame containing absorbance melting data}

\item{blank}{the blank sample}

\item{NucAcid}{A vector containing the Nucleic acid type and the sequences you are fitting.}

\item{concT}{the temperature used to calculate the NucAcid concentration. Default = 90.}

\item{Mmodel}{The molecular model you want to fit. Options: "Monomolecular.2State", "Monomolecular.3State", "Heteroduplex.2State", "Homoduplex.2State".}

\item{Tmodel}{The thermodynamic model you want to fit. Options: "VantHoff". Default = "VantHoff".}

\item{Save_results}{What results to save. Options: "all" to save PDF plots and ".csv" formated tables of parameters, "some" to save ".csv" formated tables of parameters, or "none" to save nothing.}

\item{file_prefix}{Prefix that you want on the saved files.}

\item{file_path}{Path to the directory you want to save results in.}
}
\value{
A list of data frames containing parameters from the fits and data for ploting the results with ggplot2.
}
\description{
Experimental code: not ready for use
}
