% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.F.model.R
\name{meltR.F.model}
\alias{meltR.F.model}
\title{Fit fluorescence binding isotherms to obtain thermodynamic parameters}
\usage{
meltR.F.model(
  H,
  S,
  fmax = 2,
  fmin = 0.2,
  FAM_error = 1,
  BHQ1_error = 1,
  Emission_SD = 0.05
)
}
\arguments{
\item{H}{The enthalpy of folding in kcal/mol.}

\item{S}{The entropy of folding in kcal/mol/K.}

\item{fmax}{The approximate fluorescence maximum of the fluorophore labeled strand in its single stranded state}

\item{fmin}{The approximate fluorescence minimum for the double stranded state where the fluorophore labeled strand is base paired with the quencher labeled strand.}

\item{Emission_SD}{The approximate experimental error for each emission reading.}
}
\value{
A dataframe containing the modeled data. Ready to feed into meltR.F.
}
\description{
Models data for a set of fluorescence binding isotherms given a enthalpy and entropy of folding. Assumes three replicates
of quencher labeled strand concentration =  0, 1, 10, 50, 100, 150, 200, 250, 400, 600, 800, and 1000 nM with
a constant concentration of 200 nM FAM. Generates isotherms from 20 to 80 degrees celcius with 0.5 degrees celcius
steps. Models error using the rnorm function.
}
