% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.f.R
\name{meltR.F}
\alias{meltR.F}
\title{Fit fluorescence binding isotherms to obtain thermodynamic parameters}
\usage{
meltR.F(
  df,
  Kd_error_quantile = 0.25,
  Kd_range = c(1, 250),
  Start_K = 0.1,
  Optimize_conc = TRUE,
  Low_reading = "auto",
  low_K = FALSE,
  B.conc.Tm = 250,
  Save_results = "none",
  file_prefix = "Fit",
  file_path = getwd(),
  silent = FALSE
)
}
\arguments{
\item{df}{data frame containing fluorescence binding data}

\item{Kd_error_quantile}{Quantile for uncertainty in equilibrium constants for fitting to thermodynamic models. K_error = K standard error/K. Default = 0.25 or the 25\% most accurate binding constants in the Kd_range (See "Kd_range").}

\item{Kd_range}{A custom acceptable nM Kd range to fit for your experiment. Options FALSE for no custom Kd_range or c(start, end) to set a Kd_range. Example: "Kd_range = c(5, 100)"}

\item{Start_K}{A Kd value to start non-linear regression. Default = 0.1.}

\item{Optimize_conc}{Deals with a fundamental experimental uncertainty in determination of A =fluorophore and B = Quencher concentrations in the experiment. If TRUE, meltR.f will optimize the concentration for the fluorophore labeled strand based on the shape of low temperature isotherms.}

\item{Low_reading}{Used by the concentration optimization algorithm. The isotherm, or reading, that you want to use to optimize the concentration. Default = "auto" will use the lowest temperature reading.}

\item{low_K}{Used by the concentration optimization algorithm. A low Kd value in nanomolar, that is used to find an optimum ration between A and B strands in the experiment. Default = FALSE to allow the low_K to float in the concentration optimization algorithm. It is best to use values between 0.1 and 10.}

\item{B.conc.Tm}{Only use quencher (or B strands) higher than this threshold in the 1/Tm versus lnCt fitting method, method 3}

\item{Save_results}{What results to save. Options: "all" to save PDF plots and ".csv" formatted tables of parameters, "some" to save ".csv" formatted tables of parameters, or "none" to save nothing.}

\item{file_prefix}{Prefix that you want on the saved files.}

\item{file_path}{Path to the directory you want to save results in.}

\item{silent}{Set to TRUE to run in silent mode (which does not print results in the console). Good for running in loops. Default is TRUE.}
}
\value{

}
\description{
Automates the trivial but time-consuming tasks associated with non-linear regression.
Uses two non-linear regression methods to calculate thermodynamic parameters. Method 1
fits each isotherm individually then calculates thermodynamic parameters with a Van't Hoff
plot. Method 2 calculates thermodynamic parameters with a global fit, where H and S are constant
between isotherms and the Fmax and Fmin are allowed to float. Also includes an algorithm that
optimizes the mole ratio of fluorophore labeled strands to quencher labeled strands.
}
