% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{meltr_example}
\alias{meltr_example}
\title{Get path to meltr example}
\usage{
meltr_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
A file path or a vector of file names
}
\description{
meltr comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
meltr_example()
meltr_example("mtcars.csv")
}
