% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o3a.to.MeltR.csv.R
\name{o3a.to.MeltR.csv}
\alias{o3a.to.MeltR.csv}
\title{Pulls and combines data from a single wavelength from multiple ".o3a" files}
\usage{
o3a.to.MeltR.csv(
  file,
  write = FALSE,
  Wavelength = "260",
  Remove_readings = c(),
  Pathlength = rep(1, 1000),
  Sample_names = "NULL",
  Norm_reading = FALSE
)
}
\arguments{
\item{file}{Path to a directory containing ".o3a" formatted files.}

\item{write}{Optional prefix for the ".csv" formatted file you want to write. Example is "Name". Default = FALSE to not write a file.}

\item{Wavelength}{Wavelength in nm that you want to pull out of the ".o3a" files. Default = "260".}

\item{Remove_readings}{Optional reading to remove from the data set.}

\item{Pathlength}{Pathlength in cm use to collect the data in the ".o3a" files, formatted into a vector. Example: Pathlength = c(1, 1, 0.5, 0.5, 0.1). Default = 1 cm.}

\item{Sample_names}{Optional names of the samples in each ".o3a" file, Formatted into a vector. Example: Sample_names = c("Leo", "Theo", "Betty", "Apallo") or  Sample_names = c(1:4). Default = NULL uses the ".o3a" file names as sample names}

\item{Norm_reading}{Optional reading to normalize the data to. Example: Norm_reading = 1. Default = FALSE to not normalize the data.}
}
\value{
A data frame.
}
\description{
Reads all ".o3a" files in a directory, pulls out the data from a single wavlength,
and combines the data into a data frame for plotting and subsequent data analysis.
Also includes an option to normalize the data to a reading and write the data frame
to a ".csv" file.
}
