% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.o3a.R
\name{read.o3a}
\alias{read.o3a}
\title{Reads an ".o3a" file into R}
\usage{
read.o3a(file)
}
\arguments{
\item{file}{Path to a single ".o3a" formatted file.}
}
\value{
A data frame containing the data in the "information in the ".o3a" formatted file.
}
\description{
Reads an ".o3a" file, produced by the OLIS absorbance spectrophotometer, into a R
data frame. The temperature data is in the first column of the data frame.
The absorbance for each wavelength is stored in columns that named by the absorbance
in nm. Example, "x260".
}
