% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLTrimmer.R
\name{BLTrimmer}
\alias{BLTrimmer}
\title{Auto-trim absorbance melting curves to obtain more accurate meltR.A fits}
\usage{
BLTrimmer(
  meltR.A.fit,
  n.combinations = 100,
  n.ranges = 5,
  range.step = 5,
  no.trim.range = c(0.2, 0.8),
  parallel = "none",
  n.core = 1,
  Save_results = "none",
  file_path = getwd(),
  file_prefix = "BLTrimmer"
)
}
\arguments{
\item{meltR.A.fit}{A object produced by fitting data with MeltR.A.}

\item{n.combinations}{Number of baseline combinations to test. The program will produce n.ranges^Samples combinations of baselines. It will require a large amount of computational time to test these. In general, testing 1000 combinations will produce a reliable result (plus or minus 5\% in terms on enthalpy). For an exhaustive testing, set this parameter to n.ranges^Samples.}

\item{n.ranges}{Number of trimmed baselines to generate per sample. It is not recommended to increase this parameter past 6 because of how long it will take the computer to generate all of these combinations.}

\item{range.step}{Temperature difference between each range that is generated for each sample. Default = 5 deg Celsius works well.}

\item{no.trim.range}{Determines the range where the absorbance data will not be trimmed. By default, no.trim.range = c(0.2, 0.8), meaning that the data will not be trimmed at a mode fraction double stranded greater than 0.2 and less than 0.8. Determined using the global fit from the input object created by meltR.A.}

\item{parallel}{Set to "on" if you want to use multiple cores to fit baselines. You will need to run library(doParallel) to load doParallel and its dependencies. You will also need to specify the number of cores in the n.cores argument, which should not exceed the number of cores your computer has. There will be no benefit for running in parallel mode for a single core machine. Default = "none".}

\item{n.core}{How many cores do you want to designate for this task.}

\item{Save_results}{"all" to save results to the disk or "none" to not save results to the disk.}

\item{file_path}{A path to the folder you want your results saved in.}

\item{file_prefix}{Prefix that you want on the saved files.}
}
\value{
A list of data frames containing parameters from the fits and data for plotting the results with ggplot2.
}
\description{
Automates absorbance data baseline trimming and testing. Generates random permutations of trimmed baseline combinations, fits the data using Method 1 and Method 2, compares the enthalpy difference, and chooses an optimum set of trimmed absorbance melting curves.
}
