% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLTrimmer.R
\name{BLTrimmer}
\alias{BLTrimmer}
\title{Auto-trim absorbance melting curves to obtain more accurate meltR.A fits}
\usage{
BLTrimmer(
  meltR.A.fit,
  Trim.method = "floating",
  Assess.method = 3,
  n.combinations = 1000,
  n.ranges.float = 5,
  range.step.float = 5,
  n.ranges.fixed = 40,
  range.step.fixed = 0.5,
  no.trim.range = c(0.1, 0.9),
  quantile.threshold = 0.25,
  Save_results = "none",
  file_path = getwd(),
  file_prefix = "BLTrimmer",
  memory.light = TRUE,
  Silent = FALSE
)
}
\arguments{
\item{meltR.A.fit}{A object produced by fitting data with MeltR.A.}

\item{Trim.method}{Method for trimming baselines. "fixed" to use the same baseline lengths for each curve or "floating" to use different baseline lengths for each curve. Default = "floating".}

\item{Assess.method}{Method for assessing fits from trimmed baseline. Options are integers 1, 2, and 3. 1 maximizes agreement between individual fits. 2 maximizes agreement between the average of individual fits and the average of the 1/Tm versus lnCt method. 3 takes both 1 and 2 into account. Default = 3.}

\item{n.combinations}{Number of baseline combinations to test using the float method. The program will produce n.ranges^Samples combinations of baselines. It will require a large amount of computational time to test these. In general, testing 1000 combinations will produce a reliable result (plus or minus 5\% in terms on enthalpy). For an exhaustive testing, set this parameter to n.ranges^Samples.}

\item{n.ranges.float}{Number of trimmed baselines to generate per sample using the float method. It is not recommended to increase this parameter past 6 because of how long it will take the computer to generate all of these combinations.}

\item{range.step.float}{Temperature difference between each range that is generated for each sample using the float method. Default = 5 deg Celsius works well.}

\item{n.ranges.fixed}{Number of baseline ranges for the fixed method.}

\item{no.trim.range}{Determines the range where the absorbance data will not be trimmed. By default, no.trim.range = c(0.2, 0.8), meaning that the data will not be trimmed at a mode fraction double stranded greater than 0.2 and less than 0.8. Determined using the global fit from the input object created by meltR.A.}

\item{quantile.threshold}{Threshold for assessing the best baseline combinations}

\item{Save_results}{"all" to save results to the disk or "none" to not save results to the disk.}

\item{file_path}{A path to the folder you want your results saved in.}

\item{file_prefix}{Prefix that you want on the saved files.}

\item{memory.light}{Keeps the BLTrimmer from blowing up memory usage. If TRUE, only passes the objects that the BLtrimmer needs to run to the memory. If FALSE passes a much more extensive number of statistics to the memory.}

\item{Silent}{Set to TRUE to not print results}

\item{range.step.fixed.}{Temperature difference between each range that is generated using the fixed method.}

\item{parallel}{This is argument is not exposed to end users. Contact Jacob Sieg if you are interested. to "on" if you want to use multiple cores to fit baselines. You will need to run library(doParallel) to load doParallel and its dependencies. You will also need to specify the number of cores in the n.cores argument, which should not exceed the number of cores your computer has. There will be no benefit for running in parallel mode for a single core machine. Default = "none".}

\item{n.core}{Associated with n.core and not exposed to end users. Contact Jacob Sieg if you are interested. How many cores do you want to designate for this task.}
}
\value{
A BLTrimmer fit object containing a list of data objects for advanced analysis. Set memory.light = FALSE to get the full object.
\itemize{
\item 1. Baseline.data - A data frame containing the temperature range for each sample with the dH and dS for methods 1 and 2. Also contains error between each method. This object is not returned when memory.light = TRUE.
\item 2. List.T.ranges - A list containing the temperature range for each sample and baseline combination tested. This object is not returned when memory.light = TRUE.
\item 3. List.fits - A list containing the meltR.A fit object for each baseline combination that was passed to the optimum baseline ensemble. This object is not returned when memory.light = TRUE.
\item 4. Fit.summaries - A data frame containing thermodynamic parameters from a meltR.A fit object for each baseline combination that was passed to the optimum baseline ensemble. This object is not returned when memory.light = TRUE.
\item 5. Ensemble.energies - A data frame containing the thermodynamic parameters from an analysis of the optimum ensemble energies. This object is returned when memory.light = TRUE.
\item 6. Fractional.error.between.methods - A data frame containing percent error between methods from an analysis of the optimum ensemble energies. This object is returned when memory.light = TRUE.
\item 7. System.time - A list containing the system time used by the BLTrimmer run. The first element is how long the fast analysis of many baseline combinations took. The second element is how long the fitting of the optimum baseline ensemble by meltR.A took. This object is returned when memory.light = TRUE.
}
}
\description{
Automates absorbance data baseline trimming and testing. Generates random permutations of trimmed baseline combinations, fits the data using Method 1 and Method 2, compares the enthalpy difference, and chooses an optimum set of trimmed absorbance melting curves.
}
