% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_exctinction_coef.R
\name{calc.extcoeff}
\alias{calc.extcoeff}
\title{Calculates extinction coefficents}
\usage{
calc.extcoeff(NucAcid)
}
\arguments{
\item{NucAcid}{A vector containing RNA or DNA sequences. Example "c("RNA", "ACCUUUCG", "CGAAAGGU")"}
}
\value{
A data frame containing the information in the .ct file
}
\description{
Calculates the extinnction coefficient for a DNA or RNA sequences using the
nearest neigbor method. Also calculates the sum of all the extinction
coefficients in the input vector. Nearest neighbor parameters are from:
}
\details{
Puglisi, J. D.; Tinoco, I. Absorbance Melting Curves of RNA.
Methods in Enzymology; RNA Processing Part A: General Methods;
Academic Press, 1989; Vol. 180, pp 304-325.
https://doi.org/10.1016/0076-6879(89)80108-9.
}
