% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.test.parameters}
\alias{df.test.parameters}
\title{df.test.parameters}
\format{
A data frame with 3888 rows and 9 variables:
\describe{
\item{\code{blank}}{The method used to blank the data set}
\item{\code{methods}}{The methods to use in the data set}
\item{\code{Mmodel}}{The molecular model}
\item{\code{NucAcid}}{The nucleic acid or concentrations}
\item{\code{outliers}}{The outliers to remove}
\item{\code{Tm_method}}{The method to determine the Tm in method 2}
\item{\code{Wavelength}}{The wavelength to use for extinction coefficient calculation}
\item{\code{fitTs}}{What baseline trimming to use}
\item{\code{DS}}{The data set to use}
}
}
\usage{
df.test.parameters
}
\description{
A data frame describing variables to run the meltR.A testing protocol.
}
\keyword{datasets}
