% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.a.R
\name{meltR.A}
\alias{meltR.A}
\title{Fit absorbance melting curves to obtain thermodynamic parameters}
\usage{
meltR.A(
  data_frame,
  blank = "none",
  NucAcid,
  wavelength = 260,
  concT = 90,
  fitTs = NULL,
  methods = c(TRUE, TRUE, TRUE),
  Tm_method = "nls",
  Mmodel,
  Tmodel = "VantHoff",
  Save_results = "none",
  file_prefix = "Fit",
  file_path = getwd(),
  auto.trimmed = FALSE,
  Silent = FALSE
)
}
\arguments{
\item{data_frame}{data_frame containing absorbance melting data}

\item{blank}{The blank sample for background subtraction, or a list of blanks to apply to different samples for background subtraction. "none" to turn off background subtraction. If there is a single blank in the data set, he identity of the blank, for example, blank = 1 or blank = "blank". If there are multiple blanks in the data, blank = list(c("Sample 1", "Blank 1"), c("Sample 2", "Blank 2")) and so on. Sample identifiers should be what they are in the data frame. If you need to figure out what the sample identifiers are, run unique(df$Sample), where df is the name of the R data frame you are using, in your R console.}

\item{NucAcid}{A vector containing the Nucleic acid type and the sequences you are fitting for calculating extinction coefficients. Examples: c("RNA", "UUUUUU", "AAAAAA"), c("DNA", "GCTAGC"), etc... . For a custom extinction coefficient enter "Custom" followed by the molar extinction coefficients for every nucleic acid in the sample. For example, c("Custom", 10000, 20000).}

\item{wavelength}{The wavelength you are using in the data set in nm. Options for RNA: 300, 295, 290, 285, 280, 275, 270, 265, 260, 255, 250, 245, 240, 235, and 230 nm. Options for DNA 260 nm. Most accurate at pH 7.0.}

\item{concT}{The temperature used to calculate the NucAcid concentration. Default = 90.}

\item{fitTs}{Option to only fit certain temperature ranges for melting curves. Either a vector or a list. If this is set to a vector, meltR.A will only fit temperatures in this range for all melting curves Example = c(17, 75). If set to a list of vectors, meltR.A will change what values are fit for each curve. Example, list(c(0,100), c(17,75), .... , c(0,100)). The length of this list has to be the equal to the number of samples that will be fit.}

\item{methods}{what methods do you want to use to fit data. Default = c(TRUE, TRUE, TRUE). Can be true or false. Note, method 1 must be set to TRUE or the subsequent steps will not work.}

\item{Tm_method}{either "nls" to use the Tms from the fits in Method 1, "lm" to use a numeric method based on linear regression of fraction unfolded calculated with method 1, or "polynomial" to calculate Tms using the first derivative of a polynomial that approximates each curve.}

\item{Mmodel}{The molecular model you want to fit. Options: "Monomolecular.2State", "Monomolecular.3State", "Heteroduplex.2State", "Homoduplex.2State".}

\item{Tmodel}{The thermodynamic model you want to fit. Options: "VantHoff". Default = "VantHoff".}

\item{Save_results}{What results to save. Options: "all" to save PDF plots and ".csv" formated tables of parameters, "some" to save ".csv" formated tables of parameters, or "none" to save nothing.}

\item{file_prefix}{Prefix that you want on the saved files.}

\item{file_path}{Path to the directory you want to save results in.}

\item{auto.trimmed}{Ignore this argument unless you are writting auto baseline trimmers}

\item{Silent}{TRUE to not print data in your console. Default = FALSE.}
}
\value{
A list of data frames containing parameters from the fits and data for ploting the results with ggplot2.
}
\description{
Automates the trivial but time-consuming tasks associated with non-linear regression.
Calculates extinction coefficients, subtracts out the baseline buffer readings, and
calculates the strand concentration, Ct, in each sample. Then uses three non-linear regression
methods to calculate thermodynamic parameters. Method 1 fits each melting curve individually
then reports the average H and S from all of the curves. Method 2 calculates the Tm for each
melting curve, and calculates thermodynamic parameters by fitting the relationship between Tm
and Ct. Method 3 calculates thermodynamic parameters with a global fit, where H and S are constant
between isotherms and the baslines are allowed to float. Also includes an algorithm that
optimizes the mole ratio of fluorophore labeled strands to quencher labeled strands.
}
