% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltR.f.R
\name{meltR.F}
\alias{meltR.F}
\title{Fit fluorescence binding isotherms to obtain thermodynamic parameters}
\usage{
meltR.F(
  data_frame,
  Tmodel = "VantHoff",
  K_range = FALSE,
  K_error = c(0.5, 0.5),
  Start_K = 0.1,
  Optimize_conc = TRUE,
  Low_reading = "auto",
  low_K = 0.1,
  B.conc.Tm = 500,
  Save_results = "none",
  file_prefix = "Fit",
  file_path = getwd(),
  Tm_smooth = 8
)
}
\arguments{
\item{data_frame}{data_frame containing fluorescence binding data}

\item{Tmodel}{The thermodynamic model you want to fit the data to. Options are "VantHoff" and "Kirchoff".}

\item{K_error}{Acceptable uncetainty in equillibrium constants for fitting to thermodynamic models. K_error = K standard error/K. Default = c(0.25, 0.25).}

\item{Start_K}{A K value to start non-linear regression. Default = 0.1.}

\item{Low_reading}{Used by the concentration optimization algorithm. The isotherm, or reading, that you want to use to optomize the concentration. Default = "auto" will use the lowest temperature reading.}

\item{low_K}{Used by the concentration optimization algorithm. A low K value in nanomolar, that is used to find an optimum ration between A and B strands in the experiment. Default = 0.01.}

\item{B.conc.Tm}{Only use quencher (or B strands) higher than this threshold in the 1/Tm versus lnCt fitting method, method 3}

\item{Save_results}{What results to save. Options: "all" to save PDF plots and ".csv" formated tables of parameters, "some" to save ".csv" formated tables of parameters, or "none" to save nothing.}

\item{file_prefix}{Prefix that you want on the saved files.}

\item{file_path}{Path to the directory you want to save results in.}

\item{Tm_smooth}{Number of Temperature points your want to smooth as you determine the Tm for various points. Default = 4.}

\item{K-range}{A custom acceptable nM Kd range to fit for your experiment. Options FALSE for no custom K_range or c(start, end) to set a K_range. Example: "K_range = c(5, 100)"}

\item{Optomize_B_conc}{Deals with a fundamental experimental uncertainty in determination of A =fluorophore and B = Quencher concentrations in the experiment. If TRUE, meltR.f will optimize the quencher labeled strand based on the shape of low temperature isotherms with K values < 0.1.}
}
\value{

}
\description{
Automates the trivial but time-consuming tasks associated with non-linear regression.
Uses two non-linear regression methods to calculate thermodynamic parameters. Method 1
fits each isotherm individually then calculates thermodynamic parameters with a Van't Hoff
plot. Method 2 calculates thermodynamic parameters with a global fit, where H and S are constant
between isotherms and the Fmax and Fmin are allowed to float. Also includes an algorithm that
optimizes the mole ratio of fluorophore labeled strands to quencher labeled strands.
}
